Add support for lowering xla_hlo::broadcast_in_dim op to linalg.generic.
With changes to Linalg to handle zero-rank memrefs, the lowering from
xla_lhlo::broadcast_in_dim op to linalg.generic op on buffers. can be
done in a more canonical fashion. This allows it to be adaptble to use
for xla_hlo::broadcast_in_dim op to linalg.generic op on tensors.
PiperOrigin-RevId: 296912747
Change-Id: I93bf9cfa7145bfc1f7139c38606928fa05343ed1