Use context view id to for remote tensors
The remote mirror maps were previously changed to key of the device
incarnation instead of the device pointer. However, this is not
sufficient since the incarnation can be the same for a worker when it
comes back.

To resolve this we store the context view id in the tensor handle data
and validate this in all places. Further, we change the mirror map to be
keyed off the device name so we can detect and remove entries from old
workers.

In addition, make id functions in EagerContext const and store as
reference type in RemoteTensorHandleData.

PiperOrigin-RevId: 296945105
Change-Id: If004a4ead46d4fcbe795414cfbbfa6f9998ade27