Add an eager C API for deserializing generic op attributes, TFE_OpSetAttrValueProto (like TF_SetAttrValueProto for graph building)
Also adds an experimental eager C API for serializing op attributes as generic name->value mappings

It's a bit sad that being generic requires serializing here, but I don't see a great way around it if the attributes will be used generically (e.g. to build a FunctionDef). We can add special cases that don't require serialization for fetching attributes when the type is known.

PiperOrigin-RevId: 297003316
Change-Id: Id6e65bc7a8178fbbb8a85a542bd31def08225fe6