[OpKernel] Move `output_required()` implementation from `OpKernel` to `OpKernelContext`.
The previous implementation of `OpKernel::output_required()` had a subtle bug: we could have a stateful kernel that is shared between multiple subgraphs, where each subgraph might require a different subset of the outputs. To fix this, this change moves the implementation of the outputs-required bit vector to the executor's `NodeItem`, and passes it into the kernel through `OpKernelContext`.

PiperOrigin-RevId: 297403579
Change-Id: I7e0714ae950e6c2047c2bba56bab2f1216972e21