Function calls inherit `run_all_kernels_inline` from their parent context.
The `Executor::Args::run_all_kernels_inline` flag optimizes the execution of graphs with many small kernels, and avoids potentially unbounded stack growth. This change enables functions called by a kernel to inherit this flag, which extends support for the option to larger and more complicated graphs containing function calls. It adds the flag to `OpKernelContext::Params` and `FunctionLibraryRuntime::Options`, and updates function-calling kernels to propagate it.

PiperOrigin-RevId: 298636976
Change-Id: I28263aa5a17ce7d94b84f6bb42657ce3f4b88cfa