Remove the `OpKernel::AsAsync() const` overload.
This overload was only used in `CheckNotInComputeAsync(OpKernelContext*, ...)`, which is modified to have friend access to a non-const `OpKernel*` instead. Removing the overload avoids the need to specify the same logic twice, and the risk of implementing one but not the other (a mistake I made on a recent change).

PiperOrigin-RevId: 298691746
Change-Id: Iba262216fdb6bdce4171794b08acbee393ac116b