Do not constant-fold nodes that will use ScopedAllocator.
The constant folding optimizer creates a subgraph with constant foldable nodes,
and runs them on the CPU using `GraphRunner`.  If any of the nodes in the
subgraph contain `_scoped_allocator` attribute, they are intended to be
allocated with a ScopedAllocator.  However, the `_ScopedAllocator` node is
absent from this graph, which causes the execution of this subgraph to error
out.

This change skips nodes that are marked with `_scoped_allocator` attribute in
the constant folding optimizer.

PiperOrigin-RevId: 299421744
Change-Id: Ic8e5896dbd5612d7782b9aa2e078e0128ad02f90