Add CreateTokenOp to HLO dialect.
This is to separate out the use of AfterAll op to create a HLO token out of thin air and to wait on some existing tokens.
The idea is to at some point only allow the main function to create tokens which would be passed to other computations. This would force the creators of side effecting ops to consider tokens and ordering while creating the op.

This op is exported as an AfterAll HLO with no operands. Replace the generation of AfterAll op in tf.InfeedDequeueTuple and tf.OutfeedEnqueueTuple legalization with this CreateTokenOp.

PiperOrigin-RevId: 299857684
Change-Id: I9a358f5537115ea2afac0274a38ed75f61d66b9c