Fix bug in SensitivitySpecificityBase derived metrics.
Sub-classes of `SensitivitySpecificityBase` compute the value of one statistic, given a constraint on another statistic (e.g. compute recall at specified precision). Previously the documentation stated "Computes <statistic A> at a given <statistic B> <X>.", there was no guaranteed and consistent behaviour in case the specified <statistic B> cannot be assume value <X> on the provided sample of scores and labels (e.g. required recall of 0.7, but only either 0.6 or 0.8 can be reached).

This change refines the function behaviour to "Computes best <statistic A> where <statistic B> >= <X>".

This caters to common use cases of operating binary classifiers, with a requirement to e.g. maintain a minimal precision and maximise the recall - it is important not to report a recall from an operating point that undershoots the required precision (previously the closest precision would be selected, even it if is smaller).

Because this changes (refines) the semantics of the metrics, some expected values in unittests etc. must be adapted.

PiperOrigin-RevId: 299919680
Change-Id: I496ffaef51e599ccf82222a02e2cab55ade35f1a