Improve the Reshape->QDQ passthrough pattern
Previously the reshape op in this pattern is assumed to have the same input and
output ranks. There is one new case where Reshape ops are inserted before and
after FakeQuantWithMinMaxVarsPerChannel op to bypass the quantization dimension
issue, so this patch tries to handle such a pattern. More details about the
scenario is demonstrated in the unit test.

PiperOrigin-RevId: 301264077
Change-Id: Iefdeaa58e653ad6804e0afb7a2a0c6e447d1206d