Add explicit padding to depthwise conv2d.
Explicit padding is supported with CPUs, GPUs, and XLA. No XLA code needed to be changed as the existing codepaths already worked with explicit padding. Little CPU and GPU code needed to be changed as well, as the algorithms already had to take into account padding to support SAME padding.

PiperOrigin-RevId: 301293680
Change-Id: I3021c422ac5ae2756fb273d8e5c633407f7a45df