Explicitly register dialects required in compilation pipelines
Currently, these are relying on AllPassesAndDialects target to link and register all dialects and passes. Linking only the required dialects reduces the size of linker input and binary.

This change is not re-using a static MLIR context among multiple invocations to avoid any change in behavior there and can be revisited separately.

PiperOrigin-RevId: 302344149
Change-Id: I1537ae6d1fd332fb8c9d47e5d35559c3e64ba45c