Allow adding compiler specific flags to the GPU crosstool.
For now, only support "clang", "msvc" and "unknown" (implying a variety of
gcc compatible compilers).

We hard-code the MSVC toolchain to "msvc".
The darwin toolchain is kept at "unknown", as it currently isn't tested with
any of our clang-specific flags.
The linux toolchain will use "clang" if compiling with cuda-clang, as that's
currently the only situation in which we care about using clang-specific
flags. In the future, we do optimally want to use compiler detection and
setting the compiler attribute accordingly.

PiperOrigin-RevId: 302429367
Change-Id: I3f47582e76d86d119d3fca4a6855cfc377667922