Take #2: Improve Conv2DBackpropInput to take input_sizes as a 2D shape.
When input_sizes is a 2D shape, the input batch size comes from output_grad and
the input channel size comes from the filter.

With this change, input_sizes is more likely to be a constant (e.g. even when
the batch size is variable) so tf2tensorrt is able to convert more
Conv2DBackpropInput to IDeconvolutionLayer.

Changes to tf2tensorrt will come in separate CLs.

I haven't made tf2xla support input_sizes being a 2D shape. It would error out
for now. So we disabled the test added to conv_ops_test.py for XLA.

PiperOrigin-RevId: 303217218
Change-Id: I283106657c00f49be41a74c7131bf8be787742a8