Optimize (legacy, reference-typed) `VariableOp::Compute()`.
The `cinfo_` member can be initialized at construction time, so there is no need to acquire a mutex in the `Compute()` method.

The changes in "ops_testutil.cc" update the test harness to use a non-deprecated method for creating kernels in C++ tests. The deprecated API did not set the resource manager in `OpKernelConstruction`.

Note that this op is still rather inefficient because it looks up the variable in the `ResourceMgr` every time it is invoked (for backwards compatibility reasons, relating to state invalidation via `Session::Reset()`). However, new code—and all TF2 code—should use resource variables, which avoid this codepath.

PiperOrigin-RevId: 303410876
Change-Id: I29e114bb863fe8a1f483d4ce1e034f5f13f1a116