Rolling forward "[SimplePropagatorState] Use atomic pending counts for each node in the graph."
The previous code was correct, but failed during TSAN tests because of a limitation in TSAN's happens-before relation. To address this problem, we conditionally add a read of the atomic pending count in `GetInputTensors()`, which establishes that the write to `input_tensors_` in `PropagateOutputs()` happens before the read in `PrepareInputs()`.

PiperOrigin-RevId: 304810349
Change-Id: I9058c55898079cb3ab6011513b4468ddb293f59f