Correctness fixes for dynamic_rnn/dynamic_decode under V2 control flow.
- Adds a "in_while_loop_defun" helper to control_flow_util_v2 and use it to determine whether we are within a while loop, and therefore should not cache variables.
- Use GraphOrParentsInXlaContext to determine whether XLA is in use in tf.contrib.seq2seq.dynamic_decode.

The XLA check is used to decide whether to use dynamic-sized TensorArrays or whether to size them according to maximum_iterations.

Under V2 control flow, bodies of control flow constructs are built within a FuncGraph. GetContainingXLAContext does not work in this case for determining whether XLA is in use, since the XLA context is in the context stack of the outer graph. GraphOrParentsInXlaContext both checks the current control flow context stack and the outer graph's stack.

PiperOrigin-RevId: 305380613
Change-Id: I88234bcb4963f84ea8c5d65369715e51d63f7485